refactor(formatter): split visitor traversal modules
